requirements.txt
pandas
flask
pillow
openpyxl
pycryptodome
